many: fix hysteresis bug on alarm display
Previously both the default alarm timeout and the sensor poll timeout were set
to 5 seconds and so there was no chance for the sensor value to be updated
before the alarm notification would be displayed - but then once it was
displayed the value could then have got updated to now be below the threshold,
and so the user would see a notification that was unexpectedly below the
configured threshold.

So fix this by setting the sensor poll timeout to 4 seconds, and fixing the
hysteresis value at 5 seconds so that there is a chance to update the sensor
value before the alarm notification is shown.

Also remove the ability to configure the hysteresis value as users will not need
to configure this anyway.

Signed-off-by: Alex Murray <murray.alex@gmail.com>