indicator-sensors/is-sensor.c: track error state
Previously when the error state was set for a sensor we would clear the old
state and set the new one, without checking if they were the same. But this
would result in a closing of the existing notification and showing of a new one
with the same contents.

In general it can be useful to let the user know that the sensor has encountered
and error BUT we don't want to do this every time the sensor value is tried to
be read. So check if the new error condition is different and only then take any
action, that way we won't close and redisplay the same notification.

Issue #42.

Signed-off-by: Alex Murray <murray.alex@gmail.com>